+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Group post text formatting UX [#5194](https://github.com/rokwire/illinois-app/issues/5194)
 ### Changed
 - Assistant Suggestion card to disappear only on tapping close button [#5254](https://github.com/rokwire/illinois-app/issues/5254).